+++ CHANGES.md
@@ -1,3 +1,5 @@
+# Unreleased
+  - added `Vector.swap_contents`
 
 # 1.0.0
   - first opam package