+++ CHANGES.md
@@ -1,7 +1,7 @@
 v1.2.0 2021-07-13
 -----------------
 
-* Sync MIME types with latest from Apache (@avsm)
+* Sync MIME types with latest from Apache (#22 @avsm)
 * Add `mjs` mapping to application/javascript manually (@avsm, request from @glondu)
 
 v1.1.3 2021-01-04